+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to the "keystats" extension will be documented in this file.
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
+## [1.0.0]
+
+- Super detailed keystroke logging
+- More types for keys per date / project
+- Combo mode hold while keys are pressed
+- Correctly log undefined or weird keys
+- Watch file & workspace name changes
+
 ## [0.2.0]
 
 - Log single keys instead of global